requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
idna==2.10
filestack-python==3.3.0
git+https://github.com/bambadiagne/paydunya-python
urllib3==1.26.4
gunicorn==20.1.0
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0

